Split `compile/legend` into `compile/legend/*`
